Remove old vertex struct
